Refactor: Modularize server and validate-kb scripts
Refactored src/server.ts by extracting placeholder substitution logic into a new src/placeholderSubstitutor.ts module. This simplifies the main server file and improves maintainability.

Refactored scripts/validate-kb.ts by breaking it down into three distinct modules:
- scripts/kbReport.ts: Handles report data structure and printing.
- scripts/kbFileValidator.ts: Manages validation of individual knowledge base files.
- scripts/kbPathProcessor.ts: Handles directory traversal and processing logic.
This significantly improves the modularity and readability of the knowledge base validation script.
